refactor: rename sudoku example files
